Changes after review
